Revert "Add PartitionsBuilderBase"
This reverts commit e9cbccf9876551cf0ea289c9c052e69e907648d4.
