+++ java/core/src/test/java/sleeper/core/metrics/TableMetricsTest.java
@@ -22,7 +22,6 @@ import org.junit.jupiter.api.Test;
 
 import sleeper.core.partition.Partition;
 import sleeper.core.partition.PartitionsBuilder;
-import sleeper.core.partition.PartitionsBuilderRootFirst;
 import sleeper.core.schema.Schema;
 import sleeper.core.schema.type.LongType;
 import sleeper.core.statestore.StateStore;
@@ -91,7 +90,7 @@ public class TableMetricsTest {
         void shouldReportMetricsForMultipleFilesWithDifferentRecordCounts() {
             // Given
             createInstance("test-instance");
-            PartitionsBuilderRootFirst partitionsBuilder = new PartitionsBuilder(schema)
+            PartitionsBuilder partitionsBuilder = new PartitionsBuilder(schema)
                     .singlePartition("root");
             createTable("test-table", StateStoreTestBuilder.from(partitionsBuilder)
                     .partitionFileWithRecords("root", "file1.parquet", 100L)
@@ -143,7 +142,7 @@ public class TableMetricsTest {
         void shouldReportMetricsWithTwoFilesInOnePartitionAndOneFileInOther() {
             // Given
             createInstance("test-instance");
-            PartitionsBuilderRootFirst partitionsBuilder = new PartitionsBuilder(schema)
+            PartitionsBuilder partitionsBuilder = new PartitionsBuilder(schema)
                     .rootFirst("root")
                     .splitToNewChildren("root", "L", "R", 100L);
             createTable("test-table", StateStoreTestBuilder.from(partitionsBuilder)
@@ -169,7 +168,7 @@ public class TableMetricsTest {
         void shouldReportMetricsForMultiplePartitionsWithDifferentFileCounts() {
             // Given
             createInstance("test-instance");
-            PartitionsBuilderRootFirst partitionsBuilder = new PartitionsBuilder(schema)
+            PartitionsBuilder partitionsBuilder = new PartitionsBuilder(schema)
                     .rootFirst("root")
                     .splitToNewChildren("root", "left", "right", 10L);
             createTable("test-table", StateStoreTestBuilder.from(partitionsBuilder)
@@ -195,7 +194,7 @@ public class TableMetricsTest {
         void shouldReportMetricsForMultiplePartitionsWhenOneLeafPartitionHasNoFiles() {
             // Given
             createInstance("test-instance");
-            PartitionsBuilderRootFirst partitionsBuilder = new PartitionsBuilder(schema)
+            PartitionsBuilder partitionsBuilder = new PartitionsBuilder(schema)
                     .rootFirst("root")
                     .splitToNewChildren("root", "left", "right", 10L);
             createTable("test-table", StateStoreTestBuilder.from(partitionsBuilder)
@@ -224,7 +223,7 @@ public class TableMetricsTest {
         void shouldReportMetricsWithOneFileInMultiplePartitions() {
             // Given
             createInstance("test-instance");
-            PartitionsBuilderRootFirst partitionsBuilder = new PartitionsBuilder(schema)
+            PartitionsBuilder partitionsBuilder = new PartitionsBuilder(schema)
                     .rootFirst("root")
                     .splitToNewChildren("root", "L", "R", 100L);
             createTable("test-table", StateStoreTestBuilder.from(partitionsBuilder)
@@ -249,7 +248,7 @@ public class TableMetricsTest {
         void shouldReportMetricsWithOneFileInMultiplePartitionsAndOneFileInOnePartition() {
             // Given
             createInstance("test-instance");
-            PartitionsBuilderRootFirst partitionsBuilder = new PartitionsBuilder(schema)
+            PartitionsBuilder partitionsBuilder = new PartitionsBuilder(schema)
                     .rootFirst("root")
                     .splitToNewChildren("root", "L", "R", 100L);
             createTable("test-table", StateStoreTestBuilder.from(partitionsBuilder)
