+++ java/core/src/test/java/sleeper/core/testutils/printers/PartitionsPrinterTest.java
@@ -20,7 +20,6 @@ import org.approvaltests.Approvals;
 import org.junit.jupiter.api.Test;
 
 import sleeper.core.partition.PartitionsBuilder;
-import sleeper.core.partition.PartitionsBuilderRootFirst;
 import sleeper.core.schema.Schema;
 import sleeper.core.schema.type.LongType;
 
@@ -74,9 +73,9 @@ public class PartitionsPrinterTest {
     @Test
     void shouldPrintDifferentPartitionsForOneTable() {
         // Given
-        PartitionsBuilderRootFirst partitions1 = new PartitionsBuilder(schema).rootFirst("A")
+        PartitionsBuilder partitions1 = new PartitionsBuilder(schema).rootFirst("A")
                 .splitToNewChildren("A", "B", "C", 10L);
-        PartitionsBuilderRootFirst partitions2 = new PartitionsBuilder(schema).rootFirst("1")
+        PartitionsBuilder partitions2 = new PartitionsBuilder(schema).rootFirst("1")
                 .splitToNewChildren("1", "2", "3", 20L);
 
         // When
